pyproject.toml
[project]
name = "comfyui-api-dockercpu"
version = "1.0.0"
description = "Unified ComfyUI nodes for external API integration (Replicate, Fal.ai, HuggingFace)"
readme = "README.md"
license = {text = "MIT"}
authors = [
    { name = "ComfyUI Community" }
]
requires-python = ">=3.10"
keywords = [
    "comfyui",
    "comfyui-custom-nodes",
    "replicate",
    "falai",
    "huggingface",
    "image-generation",
    "api-integration",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

dependencies = [
    "replicate>=1.0.7",
    "fal-client>=0.5.0",
    "huggingface_hub>=0.20.0",
    "soundfile>=0.12.0",
    "Pillow>=9.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/trustypangolin/ComfyUI-API-DockerCPU"
Repository = "https://github.com/trustypangolin/ComfyUI-API-DockerCPU"
Documentation = "https://github.com/trustypangolin/ComfyUI-API-DockerCPU/wiki"
"Bug Tracker" = "https://github.com/trustypangolin/ComfyUI-API-DockerCPU/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.comfy]
PublisherId = "trustypangolin"
DisplayName = "ComfyUI-API-DockerCPU"
Icon = ""

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
